WIP: implement random resolver as custom options in parallel_best
