+++ server/server.go
@@ -458,7 +458,7 @@ func createUpstreamBranches(
 		case config.UpstreamStrategyStrict:
 			upstream, err = resolver.NewStrictResolver(groupConfig, bootstrap, cfg.StartVerifyUpstream)
 		case config.UpstreamStrategyRandom:
-			upstream, err = resolver.NewRandomResolver(groupConfig, bootstrap, cfg.StartVerifyUpstream)
+			upstream, err = resolver.NewParallelBestResolver(groupConfig, bootstrap, cfg.StartVerifyUpstream)
 		}
 
 		upstreamBranches[group] = upstream
